CAPI-233(feat): add component backButton and use it in onboarding in video stream screen to reuse code
