--- src/screens/videoStream/close/index.module.css
@@ -1,6 +1,0 @@
-.buttonInner {
-	color: #fff;
-	transform: rotate(180deg);
-	font-size: 24px;
-	padding-right: 16px;
-}
