+++ src/screens/videoStream/close/index.ts
@@ -1,22 +1,12 @@
 import closeSDK from '../../../lib/closeSDK';
 import LocalStorageManager from '../../../modules/LocalStorageManager';
 import Router from '../../../modules/Router';
-import commonCSS from '../../onboardingInstructions/closeSDKButton/index.module.css';
-import backArrow from '../../onboardingInstructions/closeSDKButton/assets/backArrow.svg';
+import createBackButton from '../../../components/backButton';
 
 export default function createCloseElement(): HTMLButtonElement {
-	const button = document.createElement('button');
-	button.className = commonCSS.button;
+	const button = createBackButton();
 	button.id = 'closeButton';
 	button.setAttribute('data-test-id', 'closeButton');
-
-	const buttonArrowImage = document.createElement('img');
-	buttonArrowImage.className = commonCSS.buttonArrowImage;
-	buttonArrowImage.src = backArrow;
-
-	button.appendChild(buttonArrowImage);
-
-	button.type = 'button';
 
 	button.onclick = () => {
 		const localStorageManager = LocalStorageManager.getInstance();
